Scraper Update, RecipeList Creation
Major update changing the way that the Scraper class works.  The Scraper
now creates a master list of all recipe urls, which it uses to create
RecipeLists- an ArrayList with a category property.  The Scraper now
checks to see if data was properly accessed, and runs all eroored URLs
again at a later time to check again.  Recipe creation now verifies
whether the Recipe already exists, and appends the category to the
existing Recipe if so.